🔄 synced local 'skyvern-frontend/src/' with remote 'skyvern-frontend/src/'
<!-- ELLIPSIS_HIDDEN -->

> [!IMPORTANT]
> Enhances workflow run page with new components and routes for detailed views, improving modularity and navigation.
>
>   - **Behavior**:
>     - Adds components `WorkflowRunBlocks`, `WorkflowRunOutput`, `WorkflowRunRecording`, `WorkflowPostRunParameters` for detailed workflow run views.
>     - Updates navigation paths in `router.tsx`, `TaskDetails.tsx`, `RunWorkflowForm.tsx`, `WorkflowPage.tsx`, `Workflows.tsx` to include new routes for workflow run views.
>   - **Components**:
>     - Refactors `WorkflowRun.tsx` to use `useWorkflowRunQuery` and modularize the workflow run view.
>     - Introduces `useWorkflowRunQuery` in `useWorkflowRunQuery.ts` for fetching workflow run data.
>   - **Misc**:
>     - Removes unused imports and code in `WorkflowRun.tsx`.
>     - Adjusts pagination and navigation logic in `WorkflowRunBlocks.tsx` and `Workflows.tsx`.
>
> <sup>This description was created by </sup>[<img alt="Ellipsis" src="https://img.shields.io/badge/Ellipsis-blue?color=175173">](https://www.ellipsis.dev?ref=Skyvern-AI%2Fskyvern-cloud&utm_source=github&utm_medium=referral)<sup> for c955873cb2ca0470225762e7748e22ccd0faaf50. It will automatically update as commits are pushed.</sup>

<!-- ELLIPSIS_HIDDEN -->
