+++ skyvern-frontend/src/routes/workflows/WorkflowPage.tsx
@@ -132,14 +132,14 @@ function WorkflowPage() {
                       if (event.ctrlKey || event.metaKey) {
                         window.open(
                           window.location.origin +
-                            `/workflows/${workflowPermanentId}/${workflowRun.workflow_run_id}`,
+                            `/workflows/${workflowPermanentId}/${workflowRun.workflow_run_id}/blocks`,
                           "_blank",
                           "noopener,noreferrer",
                         );
                         return;
                       }
                       navigate(
-                        `/workflows/${workflowPermanentId}/${workflowRun.workflow_run_id}`,
+                        `/workflows/${workflowPermanentId}/${workflowRun.workflow_run_id}/blocks`,
                       );
                     }}
                     className="cursor-pointer"
